clean-up din template func
