+++ pdfType/helper.go
@@ -2,9 +2,7 @@ package pdfType
 
 import (
 	"SimpleInvoice/generator"
-	din5008A "SimpleInvoice/norms/letter/DIN-5008-a"
 	"fmt"
-	"github.com/rs/zerolog"
 	"golang.org/x/text/language"
 	"golang.org/x/text/message"
 )
@@ -60,26 +58,4 @@ func getCellWith(pdfGen *generator.PDFGenerator, percent float64) float64 {
 	maxSavePrintingWidth = maxSavePrintingWidth - pdfGen.GetMarginLeft() - pdfGen.GetMarginRight()
 
 	return (percent * maxSavePrintingWidth) / 100.0
-}
-
-func showDebugFramesDin5008A(pdfGen *generator.PDFGenerator, logger *zerolog.Logger) {
-	logger.Warn().Msg("show debug frames")
-	pdfGen.DrawLine(din5008A.HeaderStartX, din5008A.HeaderStopY, din5008A.HeaderStopX, din5008A.HeaderStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-
-	pdfGen.DrawLine(din5008A.AddressSenderTextStartX, din5008A.AddressSenderTextStartY, din5008A.AddressSenderTextStartX, din5008A.AddressSenderTextStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.AddressSenderTextStartX, din5008A.AddressSenderTextStopY, din5008A.AddressSenderTextStopX, din5008A.AddressSenderTextStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.AddressSenderTextStopX, din5008A.AddressSenderTextStartY, din5008A.AddressSenderTextStopX, din5008A.AddressSenderTextStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-
-	pdfGen.DrawLine(din5008A.AddressReceiverTextStartX, din5008A.AddressReceiverTextStartY, din5008A.AddressReceiverTextStartX, din5008A.AddressReceiverTextStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.AddressReceiverTextStartX, din5008A.AddressReceiverTextStopY, din5008A.AddressReceiverTextStopX, din5008A.AddressReceiverTextStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.AddressReceiverTextStopX, din5008A.AddressReceiverTextStartY, din5008A.AddressReceiverTextStopX, din5008A.AddressReceiverTextStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-
-	pdfGen.DrawLine(din5008A.MetaInfoStartX, din5008A.MetaInfoStartY, din5008A.MetaInfoStopX, din5008A.MetaInfoStartY, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.MetaInfoStartX, din5008A.MetaInfoStopY, din5008A.MetaInfoStopX, din5008A.MetaInfoStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.MetaInfoStartX, din5008A.MetaInfoStartY, din5008A.MetaInfoStartX, din5008A.MetaInfoStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.MetaInfoStopX, din5008A.MetaInfoStartY, din5008A.MetaInfoStopX, din5008A.MetaInfoStopY, generator.Color{R: 255, G: 64, B: 64}, 0)
-
-	pdfGen.DrawLine(din5008A.BodyStartX, din5008A.BodyStartY, din5008A.BodyStopX, din5008A.BodyStartY, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.BodyStartX, din5008A.BodyStartY, din5008A.BodyStartX, din5008A.Height-10, generator.Color{R: 255, G: 64, B: 64}, 0)
-	pdfGen.DrawLine(din5008A.BodyStopX, din5008A.BodyStartY, din5008A.BodyStopX, din5008A.Height-10, generator.Color{R: 255, G: 64, B: 64}, 0)
 }
